refactor(hashline): seal AnchorPipeline entry point
Single public entry for hashline ordering invariant
swapReversed → stripBare → stripDiff → valEdit → boundaryDups →
verifyServed → resToSpan now via anchor-pipeline only. Private
modules (served, resolve, hasher, alphabet, pure, parse, apply)
guarded by biome noRestrictedImports with src/hashline override;
HashAssign allocation seam stays public via re-exports. Migrated
src served imports and findNewEdge test to anchor-pipeline.

No behavior or payload contract change.
